Fix Replace Animation Hook

+++ Source/ACE.DatLoader/Entity/AnimationPartChange.cs
@@ -4,12 +4,12 @@ namespace ACE.DatLoader.Entity
 {
     public class AnimationPartChange : IUnpackable
     {
-        public byte PartIndex { get; private set; }
+        public ushort PartIndex { get; private set; }
         public uint PartID { get; private set; }
 
         public void Unpack(BinaryReader reader)
         {
-            PartIndex   = reader.ReadByte();
+            PartIndex   = reader.ReadUInt16();
             PartID      = reader.ReadAsDataIDOfKnownType(0x01000000);
         }
     }
